In set_iv in cipher/raw/mod.rs, set self.inner.private_iv_size.
MBEDTLS_PRIVATE(iv_size) of struct mbedtls_cipher_context_t is only
set "for ciphers with variable-length IVs" (see "cipher.h"). In
mbedtls_cipher_set_iv (in "cipher.c") ctx->iv_size is set at the
bottom of the function, but in some cases, for example GCM, there
is an early return and iv_size is never set. Therefore, if the
code in cipher/raw/mod.rs will rely on self.inner.private_iv_size
being set it must set it itself. So I have added an assignment in
set_iv, which seems to help my use case, though it is hard to know
if it will work in all cases.